backend/requirements.txt
Django>=4.0.1,<4.1
djangorestframework>=3.13.1,<3.14
django-cors-headers>=4.3.0,<4.4
psycopg2-binary>=2.9.3,<2.10
drf-spectacular>=0.22.1,<0.23
django-phonenumber-field>=7.2,<7.3
django-phonenumbers>=1.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0